libxfs: port xfs_set_inode_alloc from the kernel
To prepare to perag initialization code move to libxfs, port the
xfs_set_inode_alloc function from the kernel and make
libxfs_initialize_perag use it.  The code isn't 1:1 identical, but
AFAICT it behaves the same way.  In a future kernel release we'll
move the function into xfs_ag.c and update xfsprogs.

Signed-off-by: Darrick J. Wong <djwong@kernel.org>